* FIX: [DEBUG] fixed some errors on starting and closing application

+++ code/engine.vc2008/xrGame/inventory_upgrade_manager.cpp
@@ -36,10 +36,15 @@ Manager::Manager()
 
 Manager::~Manager()
 {
-	delete_data( m_roots );
-	delete_data( m_groups );
-	delete_data( m_upgrades );
-	delete_data( m_properties );
+	if (m_roots.size())
+		delete_data( m_roots );
+	if (m_groups.size())
+		delete_data( m_groups );
+	if (m_upgrades.size())
+		delete_data( m_upgrades );
+	if (m_properties.size())
+		delete_data( m_properties );
+
 	m_roots.clear(); // !!!!!!!!!!!!
 	m_groups.clear();
 	m_upgrades.clear();
